Update InterfaceSpec-Locomotion.md
Added collision detection comment of application based on velocity.
Why is 'z' backwards? (i have not read up on all coordiante conventions in the osvr/vrpn ecosystem).
Having movement in 2D ground plane is enough. Height of the viewpoint should be handled by a Tracker for the camera/user eye.